% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchWeight.R
\name{GetBatchWeight}
\alias{GetBatchWeight}
\title{Get gamma for a batch}
\usage{
GetBatchWeight(anchorset, celltype, beta)
}
\arguments{
\item{anchorset}{an \code{AnchorSet} object}

\item{celltype}{cell type vector for each cell}

\item{beta}{beta parameter}

\item{alpha}{alpha parameter}
}
\value{
gamma score
}
\description{
Get gamma for a batch
}
