% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{TransferData_MB}
\alias{TransferData_MB}
\title{TransferData_MB transfer data based on anchor sets of multiple baches of reference}
\usage{
TransferData_MB(
  anchorsetL,
  refdata,
  reference = NULL,
  query = NULL,
  features = NULL,
  assay = "RNA",
  integration.name = "integrated",
  weight.reduction = "pcaproject",
  l2.norm = FALSE,
  dims = 1:10,
  nn.method = "annoy",
  k.weight = 50,
  sd.weight = 1,
  eps = 0,
  n.trees = 50,
  verbose = TRUE,
  slot = "data",
  prediction.assay = FALSE,
  store.weights = TRUE
)
}
\arguments{
\item{anchorsetL}{An \code{\link{AnchorSet}} object list generated by
\code{\link{FindTransferAnchors}}}

\item{refdata}{Data to transfer. This can be specified in one of two ways:
\itemize{
\item{The reference data itself as either a vector where the names
correspond to the reference cells, or a matrix, where the column names
correspond to the reference cells.}
\item{The name of the metadata field or assay from the reference object
provided. This requires the reference parameter to be specified. If pulling
assay data in this manner, it will pull the data from the data slot. To
transfer data from other slots, please pull the data explicitly with
\code{\link{GetAssayData}} and provide that matrix here.}
}}

\item{reference}{\code{\link{Seurat}} object to use as the reference}

\item{query}{\code{\link{Seurat}} object to use as the query}

\item{features}{Features to use for dimensional reduction. If not specified,
set as variable features of the reference object which are also present in
the query.}

\item{assay}{An \code{Assay} object}

\item{integration.name}{Integration slot name}

\item{weight.reduction}{Dimension reduction to use when calculating anchor weights.
This can be either:
\itemize{
\item{A string, specifying the name of a dimension reduction present in all objects to be integrated}
\item{A vector of strings, specifying the name of a dimension reduction to use for each object to be integrated}
\item{NULL, in which case a new PCA will be calculated and used to calculate anchor weights}
}
Note that, if specified, the requested dimension reduction will only be used for calculating anchor weights in the
first merge between reference and query, as the merged object will subsequently contain more cells than was in
query, and weights will need to be calculated for all cells in the object.}

\item{l2.norm}{Perform L2 normalization on the cell embeddings after
dimensional reduction}

\item{dims}{Number of PCs to use in the weighting procedure}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann,
annoy}

\item{k.weight}{Number of neighbors to consider when weighting anchors}

\item{sd.weight}{Controls the bandwidth of the Gaussian kernel for weighting}

\item{eps}{Error bound on the neighbor finding algorithm (from
\code{\link{RANN}})}

\item{n.trees}{More trees gives higher precision when using annoy approximate
nearest neighbor search}

\item{verbose}{Display messages/progress}

\item{slot}{Slot to store the imputed data. Must be either "data" (default)
or "counts"}

\item{prediction.assay}{Return an \code{Assay} object with the prediction
scores for each class stored in the \code{data} slot.}

\item{store.weights}{Optionally store the weights matrix used for predictions
in the returned query object.}
}
\value{
If \code{query} is provided, a modified query object is returned. For
the categorical data in refdata, prediction scores are stored as Assays
(prediction.score.NAME) and two additional metadata fields: predicted.NAME
}
\description{
TransferData_MB transfer data based on anchor sets of multiple baches of reference
}
